Re-organise models to make it simpler to contribute.
