+++ uSync.Migrations/Migrators/Community/CrumpledMarkdownEditor/Models/CrumpledMarkDown.cs
@@ -1,0 +1,9 @@
+﻿namespace uSync.Migrations.Migrators.Community.CrumpledMarkdownEditor.Models
+{
+    internal class CrumpledMarkDown
+    {
+        public CrumpledMarkdownEditor? Editor { get; set; }
+        public string? UniqueId { get; set; }
+
+    }
+}
