+++ uSync.Migrations/Migrators/Community/CrumpledMarkdownEditor/Models/CrumpledMarkdownEditor.cs
@@ -1,0 +1,8 @@
+﻿namespace uSync.Migrations.Migrators.Community.CrumpledMarkdownEditor.Models
+{
+    internal class CrumpledMarkdownEditor
+    {
+        public string? Alias { get; set; }
+        public string? Content { get; set; }
+    }
+}
