+++ uSync.Migrations/Migrators/Community/ImulusUrlPicker/Models/UrlPickerTypes.cs
@@ -1,0 +1,9 @@
+﻿namespace uSync.Migrations.Migrators.Community.ImulusUrlPicker.Models
+{
+    public enum UrlPickerTypes
+    {
+        Content,
+        Media,
+        Url
+    }
+}
