+++ uSync.Migrations/Migrators/Community/NuPickers/Models/NuPickersDotNetTypeConfig.cs
@@ -1,0 +1,8 @@
+﻿namespace uSync.Migrations.Migrators.Community.NuPickers.Models
+{
+    public class NuPickersDotNetTypeConfig
+    {
+        public string? AssemblyName { get; set; }
+        public string? ClassName { get; set; }
+    }
+}
