+++ uSync.Migrations/Migrators/Community/NuPickers/Models/NuPickersEnumConfig.cs
@@ -1,0 +1,9 @@
+﻿namespace uSync.Migrations.Migrators.Community.NuPickers.Models
+{
+    public class NuPickersEnumConfig
+    {
+        public string? AssemblyName { get; set; }
+        public string? ApiController { get; set; }
+        public string? EnumName { get; set; }
+    }
+}
