+++ uSync.Migrations/Migrators/Community/NuPickers/Models/NuPickersSqlConfig.cs
@@ -1,0 +1,8 @@
+﻿namespace uSync.Migrations.Migrators.Community.NuPickers.Models
+{
+    public class NuPickersSqlConfig
+    {
+        public string? Query { get; set; }
+        public string? ConnectionString { get; set; }
+    }
+}
