+++ uSync.Migrations/Migrators/Community/NuPickers/Models/NuPickersXmlConfig.cs
@@ -1,0 +1,11 @@
+﻿namespace uSync.Migrations.Migrators.Community.NuPickers.Models
+{
+    public class NuPickersXmlConfig
+    {
+        public string? ApiController { get; set; }
+        public string? XmlData { get; set; }
+        public string? XPath { get; set; }
+        public string? KeyXPath { get; set; }
+        public string? LabelXPath { get; set; }
+    }
+}
