+++ uSync.Migrations/Migrators/Community/NuPickers/NuPickersEnumDropDownPickerToContentmentDataList.cs
@@ -4,8 +4,8 @@ using Newtonsoft.Json.Linq;
 using Umbraco.Extensions;
 using uSync.Migrations.Context;
 using uSync.Migrations.Extensions;
+using uSync.Migrations.Migrators.Community.NuPickers.Models;
 using uSync.Migrations.Migrators.Models;
-using uSync.Migrations.Migrators.Models.NuPickers;
 
 namespace uSync.Migrations.Migrators.Community;
