+++ uSync.Migrations/Migrators/Community/NuPickers/NuPickersXPathCheckboxPickerToContentmentDataList .cs
@@ -3,8 +3,8 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using uSync.Migrations.Context;
 using uSync.Migrations.Extensions;
+using uSync.Migrations.Migrators.Community.NuPickers.Models;
 using uSync.Migrations.Migrators.Models;
-using uSync.Migrations.Migrators.Models.NuPickers;
 
 namespace uSync.Migrations.Migrators.Community
 {
