--- uSync.Migrations/Migrators/Models/CrumpledMarkDown.cs
@@ -1,9 +1,0 @@
-﻿namespace uSync.Migrations.Migrators.Models
-{
-    internal class CrumpledMarkDown
-    {
-        public CrumpledMarkdownEditor? Editor { get; set; }
-        public string? UniqueId { get; set; }
-
-    }
-}
