--- uSync.Migrations/Migrators/Models/CrumpledMarkdownEditor.cs
@@ -1,8 +1,0 @@
-﻿namespace uSync.Migrations.Migrators.Models
-{
-    internal class CrumpledMarkdownEditor
-    {
-        public string? Alias { get; set; }
-        public string? Content { get; set; }
-    }
-}
