--- uSync.Migrations/Migrators/Models/NuPickers/NuPickersDotNetTypeConfig.cs
@@ -1,8 +1,0 @@
-﻿namespace uSync.Migrations.Migrators.Models.NuPickers
-{
-    public class NuPickersDotNetTypeConfig
-    {
-        public string? AssemblyName { get; set; }
-        public string? ClassName { get; set; }
-    }
-}
