--- uSync.Migrations/Migrators/Models/NuPickers/NuPickersEnumConfig.cs
@@ -1,9 +1,0 @@
-﻿namespace uSync.Migrations.Migrators.Models.NuPickers
-{
-    public class NuPickersEnumConfig
-    {
-        public string? AssemblyName { get; set; }
-        public string? ApiController { get; set; }
-        public string? EnumName { get; set; }
-    }
-}
