--- uSync.Migrations/Migrators/Models/NuPickers/NuPickersSqlConfig.cs
@@ -1,8 +1,0 @@
-﻿namespace uSync.Migrations.Migrators.Models.NuPickers
-{
-    public class NuPickersSqlConfig
-    {
-        public string? Query { get; set; }
-        public string? ConnectionString { get; set; }
-    }
-}
