--- uSync.Migrations/Migrators/Models/NuPickers/NuPickersXmlConfig.cs
@@ -1,11 +1,0 @@
-﻿namespace uSync.Migrations.Migrators.Models.NuPickers
-{
-    public class NuPickersXmlConfig
-    {
-        public string? ApiController { get; set; }
-        public string? XmlData { get; set; }
-        public string? XPath { get; set; }
-        public string? KeyXPath { get; set; }
-        public string? LabelXPath { get; set; }
-    }
-}
