--- uSync.Migrations/Migrators/Models/UrlPicker/UrlPicker.cs
@@ -1,11 +1,0 @@
-﻿using Newtonsoft.Json.Serialization;
-using Newtonsoft.Json;
-
-namespace uSync.Migrations.Migrators.Models.UrlPicker
-{
-    [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
-    public class UrlPicker
-    {
-        public UrlPickerValue? Value { get; set; }
-    }
-}
