--- uSync.Migrations/Migrators/Models/UrlPicker/UrlPickerTypeData.cs
@@ -1,14 +1,0 @@
-﻿using Newtonsoft.Json.Serialization;
-using Newtonsoft.Json;
-
-namespace uSync.Migrations.Migrators.Models.UrlPicker
-{
-    [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
-    public class UrlPickerTypeData
-    {
-        public string? Url { get; set; }
-        public string? ContentId { get; set; }
-        public string? MediaId { get; set; }
-
-    }
-}
