--- uSync.Migrations/Migrators/Models/UrlPicker/UrlPickerTypes.cs
@@ -1,9 +1,0 @@
-﻿namespace uSync.Migrations.Migrators.Models.UrlPicker
-{
-    public enum UrlPickerTypes
-    {
-        Content,
-        Media,
-        Url
-    }
-}
